AI_RAG_CHATBOT/requirements.txt
streamlit
flask
requests
python-dotenv

openai
langchain
langchain-community
langchain-openai

crawl4ai
faiss-cpu

